start experimenting with processor grids

+++ include/gran/bus/simple_bus.h
@@ -8,6 +8,6 @@
 
 struct component *create_simple_bus();
 stat simple_bus_add(struct component *bus, struct component *component,
-                    uintptr_t addr, size_t size);
+                    uint64_t addr, uint64_t size);
 
 #endif /* GRAN_SIMPLE_BUS_H */
